+++ apps/api/CHANGELOG.md
@@ -1,5 +1,24 @@
 # api
 
+## 1.12.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [2cde454]
+  - @cgr/schema@1.6.0-beta.0
+
+## 1.12.0-beta.1
+
+### Minor Changes
+
+- 144b442: finalize ai moderation
+
+## 1.12.0-beta.0
+
+### Minor Changes
+
+- 5db09ca: add openai moderation to approve review
+
 ## 1.11.0
 
 ### Minor Changes